pin setzen

+++ RPINeopixel/Neopixel.h
@@ -20,7 +20,7 @@ public:
 
     void              begin(void);
     void              show(void);
-    void               setPin(uint16_t p);
+    void              setPin(uint8_t p);
     void              setPixelColor(uint16_t n, uint8_t r, uint8_t g, uint8_t b);
     void              setPixelColor(uint16_t n, uint8_t r, uint8_t g, uint8_t b,
                                     uint8_t w);
@@ -34,7 +34,7 @@ public:
     Neopixel(int n, int pin);
 
 private:
-    int pin;
+    uint8_t pin;
     int numpixels;
     uint32_t* buffer;
 };
